[runtime] Fix handler block
Instead of hijacking the return address when suspending a thread in a finally block for an abort, do an explicit check in managed code when we are returning from the handler in order to see if we need to abort. Hijacking the return address is arch specific and it doesn't work if we are suspended before it is initialized in handler start.

Fixes mono/mono#5798


Commit migrated from https://github.com/mono/mono/commit/ba95f34c0a23e98f2e08f0b2a16df36a66fec006