Add missing null checks to TypeReferenceExtensions
Cecil's MetadataResolver returns null if it cannot resolve something.

This requires the callers, like TypeReferenceExtensions, to do null
checks, otherwise a  `NullReferenceException` can happens in the linker
code and it's not possible to give a useful error message to the
developer.

This is what's happening in
https://github.com/xamarin/xamarin-macios/issues/3215 (with test case)
and likely in
https://github.com/xamarin/xamarin-macios/issues/3210 (without test case)


Commit migrated from https://github.com/dotnet/linker/commit/3718965d68e5980ae602501b7c4502b9eda18186