[marshal] For generic method icalls, handle generic param arguments when using coop handles
Suppose we wanted to use coop handles to implement the following icall:

   [MethodImpl (InternalCall)]
   extern void GetGenericValueImpl<T> (int idx, T out value);

The question is: how should we pass the argument "value" to the C code?

If it was a reference type like "string" , we would use a coop handle:
    void ves_icall_GetGenericValueImpl_string (gint32 idx, MonoStringHandleOut value, MonoError *error);

But if it was a valuetype, we would instead pin the memory with an interior
handle and pass the raw address:
    void ves_icall_GetGenericValueImpl_int32 (gint32 idx, gint32 *value, MonoError *error);

But we don't have two different implementation on the C side - there's just one
icall - we have to choose a single mechanism.

So the mechanism we choose is to treat all the args like valuetypes - pin the
memory in the wrapper and pass a raw pointer to the C side.

The other troubling possibility is if someone writes:
   [MethodImpl (InternalCall)]
   extern void BadInternalCall<T> (Something s, T bad);

Now we really don't have any options.  So we just assert that noone ever writes
such a call.


Commit migrated from https://github.com/mono/mono/commit/52514effbf19992edc323ff9d7bf21766bc50b2c