Tolerate assembly resolution errors (dotnet/linker#260)
* Tolerate assembly resolution errors

Previously, we swallowed AssemblyResolutionException in a few places
in various steps. We need an option to be generally more tolerant of
AssemblyResolutionExceptions, because there can be cases during
portable publish in which referenced assemblies are not present during
publish because they're part of the shared framework.

Instead of handling AssemblyResolutionException in every place in the
linker that tries to resolve methods/types/etc, this change adds an
option that enables the resolver to return null when an assembly is
not found. The null result is also cached to avoid multiple lookups.

Only a few places in the linker need to be modified to handle null
results, and the existing places where we used to catch
AssemblyResolutionException now only need to deal with a null result
instead. The downside of this approach is that these places cannot
distinguish between types not found in an existing assembly, and
assemblies that were not found.

The new behavior is grouped under the option --skip-unresolved, which
already handles unresolved types and methods in some cases. When this
option is enabled, a warning is logged, and no exception is thrown.

* Attempt to fix test failure

Introduce a SkipUnresolved attribute

* Address PR feedback

- follow coding conventions
- use LinkContext::LogMessage instead of Console.WriteLine
- delay initialization of unresolved assembly cache to catch block
- keep AssemblyResolver.IgnoreUnresolved independent of
  LinkContext.IgnoreUnresolved
- remove redundant initialization

* Handle null _unresolvedAssemblies in Dispose method


Commit migrated from https://github.com/dotnet/linker/commit/2dee7d0dddd7b793a8536a5a42d86541b0af3c46